requirements.txt
appnope==0.1.0
coverage==4.2
decorator==4.0.10
Django==1.10.3
ipython==5.1.0
ipython-genutils==0.1.0
jsonschema==2.5.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.1.3
PyYAML==3.12
simplegeneric==0.8.1
six==1.10.0
trafaret==0.7.5
traitlets==4.3.1
wcwidth==0.1.7
